Define repo parsers as strategies
